--- Sources/App/Features/Presentations/Models/PresentationResponse+V2.swift
@@ -1,11 +1,0 @@
-import Foundation
-import Vapor
-
-struct PresentationResponseV2: Content {
-    let id: UUID
-    let title: String
-    let synopsis: String
-    var speakers: [SpeakerResponse] = []
-    let slidoURL: String?
-    let videoURL: String?
-}
