--- Sources/App/Features/Schedule/Models/ScheduleResponse+V2.swift
@@ -1,8 +1,0 @@
-import Foundation
-import Vapor
-
-struct ScheduleResponseV2: Content {
-    let event: EventResponse
-    let events: [EventResponse]
-    let slots: [SlotResponseV2]
-}
