+++ Sources/App/Features/Schedule/Models/ScheduleResponse.swift
@@ -3,5 +3,6 @@ import Vapor
 
 struct ScheduleResponse: Content {
     let event: EventResponse
+    let events: [EventResponse]
     let slots: [SlotResponse]
 }
