+++ Sources/App/Features/Slots/Models/SlotResponse+V2.swift
@@ -1,11 +1,11 @@
 import Foundation
 import Vapor
 
-struct SlotResponseV2: Content {
+struct SlotResponse: Content {
     let id: UUID
     let startTime: String
     let date: Date?
     let duration: Double
-    let presentation: PresentationResponseV2?
+    let presentation: PresentationResponse?
     let activity: ActivityResponse?
 }
